split handleQueryEvent to small functions
